+++ lib/plugins/index.js
@@ -39,7 +39,7 @@ module.exports = [
   require('./aws/package/compile/events/apiGateway/index.js'),
   require('./aws/package/compile/events/websockets/index.js'),
   require('./aws/package/compile/events/sns.js'),
-  require('./aws/package/compile/events/stream/index.js'),
+  require('./aws/package/compile/events/stream.js'),
   require('./aws/package/compile/events/msk/index.js'),
   require('./aws/package/compile/events/alb/index.js'),
   require('./aws/package/compile/events/alexaSkill.js'),
